Separate the VM into multiple files
This should improve the build performance, as usually only one file will
be modified at a time.
