+++ mu/InstructionProcessor.cpp
@@ -1,0 +1,52 @@
+#include <doctest.h>
+
+#include <stdexcept>
+using std::logic_error;
+
+#include <fmt/core.h>
+using fmt::format;
+
+#include "Bytecode.hpp"
+#include "InstructionProcessor.hpp"
+#include "Interpreter.hpp"
+#include "ValueStack.hpp"
+
+// == Instruction Processor ==
+//
+// The instruction processor identifies each instruction and calls the proper
+// implemenatation for that instruction.
+
+void Process(span<Instruction> instructions) {
+  for (auto& ins : instructions) {
+    if (ins.opCode == OpCode::Push)
+      Push(ins.argument);
+    else if (ins.opCode == OpCode::Pop)
+      Pop();
+    else if (ins.opCode == OpCode::Add)
+      Add();
+    else if (ins.opCode == OpCode::Subtract)
+      Subtract();
+    else
+      throw logic_error(format("Unexpected opcode: {}", (int)ins.opCode));
+  }
+}
+
+TEST_CASE("Verify instruction processing behavior") {
+  SUBCASE("Push two values and add") {
+    Instruction instructions[] = {
+        {OpCode::Push, 2}, {OpCode::Push, 3}, {OpCode::Add}};
+    Process(instructions);
+    CHECK(Pop() == 5);
+  }
+
+  SUBCASE("Push three values, pop one and subtract") {
+    // This should do 2 - 3
+    Instruction instructions[] = {{OpCode::Push, 2},
+                                  {OpCode::Push, 3},
+                                  {OpCode::Push, 8},
+                                  {OpCode::Pop},
+                                  {OpCode::Subtract}};
+    Process(instructions);
+    CHECK(Pop() == -1);
+  }
+}
