+++ mu/InstructionProcessor.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <span>
+using std::span;
+
+#include "Bytecode.hpp"
+
+void Process(span<Instruction> instructions);
