+++ mu/Interpreter.hpp
@@ -1,0 +1,4 @@
+#pragma once
+
+void Add();
+void Subtract();
