+++ mu/ValueStack.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+void Push(int value);
+int Pop();
+int StackSize();
